:recycle: Preview 部分をコンポーネントとして切り出す
